Use uint8_t for lookup indices and escape strings in codegen
- Change lookup function signatures from uint16_t to uint8_t since
  indices are stored in uint8_t fields (max 255 per category)
- Use cpp_string_escape() for string table entries to prevent
  compilation errors from special characters in user-provided strings

+++ esphome/core/entity_base.cpp
@@ -46,9 +46,9 @@ void EntityBase::set_name(const char *name, uint32_t object_id_hash) {
 }
 
 // Weak default lookup functions — overridden by generated code in main.cpp
-__attribute__((weak)) const char *entity_device_class_lookup(uint16_t) { return ""; }
-__attribute__((weak)) const char *entity_uom_lookup(uint16_t) { return ""; }
-__attribute__((weak)) const char *entity_icon_lookup(uint16_t) { return ""; }
+__attribute__((weak)) const char *entity_device_class_lookup(uint8_t) { return ""; }
+__attribute__((weak)) const char *entity_uom_lookup(uint8_t) { return ""; }
+__attribute__((weak)) const char *entity_icon_lookup(uint8_t) { return ""; }
 
 // Entity device class (from index)
 StringRef EntityBase::get_device_class_ref() const {
